JAVA-2416 Move/rename module spring-dispatcher-servlet
- Moved all the contents from spring-dispatcher-servlet to spring-mvc-basics
