--- spring-dispatcher-servlet/README.md
@@ -1,7 +1,0 @@
-## Spring DispatcherServlet
-
-This module contains articles about Spring DispatcherServlet
-
-## Relevant articles:
-
-- [An Intro to the Spring DispatcherServlet](https://www.baeldung.com/spring-dispatcherservlet)
