--- spring-dispatcher-servlet/src/main/java/com/baeldung/springdispatcherservlet/configuration/AppConfig.java
@@ -1,83 +1,0 @@
-package com.baeldung.springdispatcherservlet.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.ui.context.support.ResourceBundleThemeSource;
-import org.springframework.web.multipart.commons.CommonsMultipartResolver;
-import org.springframework.web.servlet.config.annotation.*;
-import org.springframework.web.servlet.resource.PathResourceResolver;
-import org.springframework.web.servlet.theme.CookieThemeResolver;
-import org.springframework.web.servlet.theme.ThemeChangeInterceptor;
-import org.springframework.web.servlet.view.JstlView;
-import org.springframework.web.servlet.view.UrlBasedViewResolver;
-
-import java.io.IOException;
-
-@Configuration
-@EnableWebMvc
-@ComponentScan("com.baeldung.springdispatcherservlet")
-public class AppConfig implements WebMvcConfigurer {
-
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/").setViewName("index");
-    }
-
-    /** Multipart file uploading configuratioin */
-    @Bean
-    public CommonsMultipartResolver multipartResolver() throws IOException {
-        CommonsMultipartResolver resolver = new CommonsMultipartResolver();
-        resolver.setMaxUploadSize(10000000);
-        return resolver;
-    }
-
-    /** View resolver for JSP */
-    @Bean
-    public UrlBasedViewResolver viewResolver() {
-        UrlBasedViewResolver resolver = new UrlBasedViewResolver();
-        resolver.setPrefix("/WEB-INF/jsp/");
-        resolver.setSuffix(".jsp");
-        resolver.setViewClass(JstlView.class);
-        return resolver;
-    }
-
-    /** Static resource locations including themes*/
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/resources/**/*")
-                .addResourceLocations("/", "/resources/")
-                .setCachePeriod(3600)
-                .resourceChain(true)
-                .addResolver(new PathResourceResolver());
-    }
-
-    /** BEGIN theme configuration */
-    @Bean
-    public ResourceBundleThemeSource themeSource(){
-        ResourceBundleThemeSource themeSource = new ResourceBundleThemeSource();
-        themeSource.setDefaultEncoding("UTF-8");
-        themeSource.setBasenamePrefix("themes.");
-        return themeSource;
-    }
-
-    @Bean
-    public CookieThemeResolver themeResolver(){
-        CookieThemeResolver resolver = new CookieThemeResolver();
-        resolver.setDefaultThemeName("default");
-        resolver.setCookieName("example-theme-cookie");
-        return resolver;
-    }
-
-    @Bean
-    public ThemeChangeInterceptor themeChangeInterceptor(){
-        ThemeChangeInterceptor interceptor = new ThemeChangeInterceptor();
-        interceptor.setParamName("theme");
-        return interceptor;
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(themeChangeInterceptor());
-    }
-    /** END theme configuration */
-}
