--- spring-dispatcher-servlet/src/main/java/com/baeldung/springdispatcherservlet/configuration/WebAppInitializer.java
@@ -1,23 +1,0 @@
-package com.baeldung.springdispatcherservlet.configuration;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
-
-public class WebAppInitializer implements WebApplicationInitializer {
-
-	@Override
-	public void onStartup(ServletContext container) throws ServletException {
-		AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-		context.register(AppConfig.class);
-		context.setServletContext(container);
-
-		ServletRegistration.Dynamic servlet = container.addServlet("dispatcher", new DispatcherServlet(context));
-		servlet.setLoadOnStartup(1);
-		servlet.addMapping("/");
-	}
-}
