--- spring-dispatcher-servlet/src/main/webapp/resources/css/default.css
@@ -1,4 +1,0 @@
-h2 {
-    color: green;
-    font-weight: 400;
-}
