--- spring-dispatcher-servlet/src/main/webapp/resources/css/example.css
@@ -1,4 +1,0 @@
-h2 {
-    color: red;
-    font-weight: 700;
-}
