+++ spring-mvc-basics/README.md
@@ -8,6 +8,7 @@ The "REST With Spring" Classes: http://bit.ly/restwithspring
 
 ### Relevant Articles: 
 - [Spring MVC Tutorial](https://www.baeldung.com/spring-mvc-tutorial)
+- [An Intro to the Spring DispatcherServlet] (https://www.baeldung.com/spring-dispatcherservlet)
 - [The Spring @Controller and @RestController Annotations](https://www.baeldung.com/spring-controller-vs-restcontroller)
 - [A Guide to the ViewResolver in Spring MVC](https://www.baeldung.com/spring-mvc-view-resolver-tutorial)
 - [Guide to Spring Handler Mappings](https://www.baeldung.com/spring-handler-mappings)
