+++ spring-mvc-basics/pom.xml
@@ -24,6 +24,11 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-validation</artifactId>
         </dependency>
+        <dependency>
+            <groupId>commons-fileupload</groupId>
+            <artifactId>commons-fileupload</artifactId>
+            <version>${commons-fileupload.version}</version>
+        </dependency>        
         <!-- to enable JSP -->
         <dependency>
             <groupId>org.apache.tomcat.embed</groupId>
