+++ spring-mvc-basics/src/main/java/com/baeldung/model/User.java
@@ -1,4 +1,4 @@
-package com.baeldung.springdispatcherservlet.domain;
+package com.baeldung.model;
 
 
 public class User {
