+++ spring-mvc-basics/src/main/java/com/baeldung/spring/web/config/WebConfig.java
@@ -1,13 +1,22 @@
 package com.baeldung.spring.web.config;
 
+import java.io.IOException;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.MediaType;
+import org.springframework.ui.context.support.ResourceBundleThemeSource;
+import org.springframework.web.multipart.commons.CommonsMultipartResolver;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.ContentNegotiationConfigurer;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.resource.PathResourceResolver;
+import org.springframework.web.servlet.theme.CookieThemeResolver;
+import org.springframework.web.servlet.theme.ThemeChangeInterceptor;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 import org.springframework.web.servlet.view.JstlView;
 import org.springframework.web.servlet.view.ResourceBundleViewResolver;
@@ -24,6 +33,14 @@ public class WebConfig implements WebMvcConfigurer {
             .setViewName("index");
     }
 
+    /** Multipart file uploading configuratioin */
+    @Bean
+    public CommonsMultipartResolver multipartResolver() throws IOException {
+        CommonsMultipartResolver resolver = new CommonsMultipartResolver();
+        resolver.setMaxUploadSize(10000000);
+        return resolver;
+    }
+
     @Bean
     public ViewResolver viewResolver() {
         final InternalResourceViewResolver bean = new InternalResourceViewResolver();
@@ -33,6 +50,47 @@ public class WebConfig implements WebMvcConfigurer {
         bean.setOrder(2);
         return bean;
     }
+
+    /** Static resource locations including themes*/
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/resources/**/*")
+            .addResourceLocations("/", "/resources/")
+            .setCachePeriod(3600)
+            .resourceChain(true)
+            .addResolver(new PathResourceResolver());
+    }
+
+    /** BEGIN theme configuration */
+    @Bean
+    public ResourceBundleThemeSource themeSource() {
+        ResourceBundleThemeSource themeSource = new ResourceBundleThemeSource();
+        themeSource.setDefaultEncoding("UTF-8");
+        themeSource.setBasenamePrefix("themes.");
+        return themeSource;
+    }
+
+    @Bean
+    public CookieThemeResolver themeResolver() {
+        CookieThemeResolver resolver = new CookieThemeResolver();
+        resolver.setDefaultThemeName("default");
+        resolver.setCookieName("example-theme-cookie");
+        return resolver;
+    }
+
+    @Bean
+    public ThemeChangeInterceptor themeChangeInterceptor() {
+        ThemeChangeInterceptor interceptor = new ThemeChangeInterceptor();
+        interceptor.setParamName("theme");
+        return interceptor;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(themeChangeInterceptor());
+    }
+
+    /** END theme configuration */
 
     @Bean
     public ViewResolver resourceBundleViewResolver() {
