+++ spring-mvc-basics/src/main/java/com/baeldung/web/controller/MultipartController.java
@@ -1,4 +1,4 @@
-package com.baeldung.springdispatcherservlet.controller;
+package com.baeldung.web.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
