+++ spring-mvc-basics/src/main/resources/themes/default.properties
