+++ spring-mvc-basics/src/main/resources/themes/example.properties
